+++ cmd/chain-maind/app/app.go
@@ -37,6 +37,7 @@ import (
 	govv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 
 	"github.com/crypto-org-chain/chain-main/v4/app"
+	appcfg "github.com/crypto-org-chain/chain-main/v4/app/config"
 	"github.com/crypto-org-chain/chain-main/v4/app/params"
 	"github.com/crypto-org-chain/chain-main/v4/config"
 	chainmaincli "github.com/crypto-org-chain/chain-main/v4/x/chainmain/client/cli"
@@ -99,6 +100,8 @@ func initAppConfig() (string, interface{}) {
 
 	type CustomAppConfig struct {
 		serverconfig.Config
+
+		MemIAVL appcfg.MemIAVLConfig `mapstructure:"memiavl"`
 	}
 
 	// Optionally allow the chain developer to overwrite the SDK's default
@@ -108,10 +111,11 @@ func initAppConfig() (string, interface{}) {
 	srvCfg.GRPCWeb.Address = "127.0.0.1:9091"
 
 	customAppConfig := CustomAppConfig{
-		Config: *srvCfg,
+		Config:  *srvCfg,
+		MemIAVL: appcfg.DefaultMemIAVLConfig(),
 	}
 
-	return serverconfig.DefaultConfigTemplate, customAppConfig
+	return serverconfig.DefaultConfigTemplate + appcfg.DefaultConfigTemplate, customAppConfig
 }
 
 func initRootCmd(rootCmd *cobra.Command, encodingConfig params.EncodingConfig) {